light_os/requirements.txt
pillow>=9.0.0
psutil>=5.9.0
requests>=2.28.0
python-dotenv>=0.20.0
